+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @glissade/eslint-plugin
 
+## 0.49.0-pre.0
+
 ## 0.48.0
 
 ## 0.48.0-pre.2